// src/components/Popover/types.ts
// 弹出框方向: 上、上左、上右、左、左上、左下、右、右上、右下、下、下左、下右
export type Position =
  | 'bottomLeft'
  | 'bottom'
  | 'bottomRight'
  | 'topLeft'
  | 'top'
  | 'topRight'
  | 'leftTop'
  | 'left'
  | 'leftBottom'
  | 'rightTop'
  | 'right'
  | 'rightBottom';